clear changes

+++ src/commands/utils/clear.js
@@ -9,39 +9,31 @@ module.exports = {
   run: async (client, message, args) => {
     const msgCount = parseInt(args[0], 10);
 
-    if (!msgCount || msgCount > 99)
-      return message.channel.send(`:x: | ${message.author}`);
+    if (!msgCount || msgCount < 1 ||msgCount > 99)
+      return message.channel.send(`:x: | ${message.author} Insira um numero entre 1 - 99.`);
 
     const cleaningMsg = await message.channel.messages.fetch({
       limit: msgCount + 1,
-    });
+    })
     message.channel.bulkDelete(cleaningMsg);
 
     let clearMsg = new Discord.MessageEmbed()
       .setColor("RANDOM")
       .setDescription(`** Limpando Mensagens... 🧹**`)
-      .setFooter({
-        text: "By Dino ❤️",
-        iconURL: `${client.user.displayAvatarURL({ dynamic: true })}`,
-      })
+      .setFooter({text: `By Dino ❤️`, iconURL: `${client.user.displayAvatarURL({ dynamic: true })}`})
       .setTimestamp(Date.now());
 
-    let embed = await message
-      .reply({ content: `${message.author}`, embeds: [clearMsg] })
-      .then((msg) => {
-        setTimeout(function () {
-          let clearMsg = new Discord.MessageEmbed()
-            .setColor("RANDOM")
-            .setDescription(
-              `✅ | ${message.author} apagou \`${cleaningMsg.size()}\` mensagens!`
-            )
-            .setFooter({
-              text: "By Dino ❤️",
-              iconURL: `${client.user.displayAvatarURL({ dynamic: true })}`,
-            })
-            .setTimestamp(Date.now());
-          msg.edit({ content: `${message.author}`, embeds: [clearMsg] });
-        }, 2000);
-      });
-  },
-};
+    let clearEmbed = await message.reply({ embeds: [clearMsg] }).then(msg => {
+      setTimeout( () => {
+        let clearedMsg = new Discord.MessageEmbed()
+          .setColor("RANDOM")
+          .setDescription(`✅ | ${message.author} apagou \`${msgCount}\` mensagens!`)
+          .setFooter({ text: `By Dino ❤️`, iconURL: `${client.user.displayAvatarURL({ dynamic: true })}` })
+          .setTimestamp(Date.now());
+  
+        msg.edit({ embeds: [clearedMsg] })
+      }, 2000)
+    });
+
+  }
+}
